Move clean_overscan_vector into core/procimg.py
Relocate the overscan cleaning helper from the Binospec spectrograph
module to core/procimg.py so it can be reused by other spectrographs,
as suggested in pypeit/PypeIt#2101 (review comment r3042223628).
